% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_package_help_menu_roxygen.R
\name{generate_package_help_menu_roxygen}
\alias{generate_package_help_menu_roxygen}
\title{Automatically generate a package help menu}
\usage{
generate_package_help_menu_roxygen(package_name, man_dir = "man",
  R_dir = "R")
}
\arguments{
\item{man_dir}{The path to the /man/ directory where .Rd files can be listed.
Default assumes the current working directory is the root of the package
source, and specifies the /man/ subdirectory.}

\item{R_dir}{The path to the /R/ directory where <package>.R can be saved. Default
assumes the current working directory is the root of the package source,
and specifies the /R/ subdirectory.}
}
\description{
Based on .Rd files found in the /man/ folder of the package source,
this function will generate a file <package>.R with roxygen notation that
links to all files found in the /man/ folder. The effect is that package users
will get a menu to all help documentation in the package by typing ?<package> \cr
This function is not exported because it is not supposed to be run by package users,
only by package developers. Call it with <package>:::generate_package_help_menu_roxygen()
}
\examples{

# start by opening the package project
generate_package_help_menu_roxygen()
# then do ctrl+shift+B or re-install the package some other way
}
