% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp.R
\docType{package}
\encoding{UTF-8}
\name{sftp}
\alias{sftp}
\alias{sftp-package}
\title{sftp package help menu}
\description{
This help menu was automatically generated using
the make_help_menu function available on
\url{https://github.com/stenevang/make_help_menu/}
}
\details{
\code{\link{sftp}}\cr
\code{\link{sftp_changedir}}\cr
\code{\link{sftp_connect}}\cr
\code{\link{sftp_delete}}\cr
\code{\link{sftp_download}}\cr
\code{\link{sftp_list}}\cr
\code{\link{sftp_listdirs}}\cr
\code{\link{sftp_listfiles}}\cr
\code{\link{sftp_makedir}}\cr
\code{\link{sftp_removedir}}\cr
\code{\link{sftp_rename}}\cr
\code{\link{sftp_upload}}\cr
}
