% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp_functions.R
\name{sftp_connect}
\alias{sftp_connect}
\title{Create a list object with SFTP connection details}
\usage{
sftp_connect(server = "", folder = "", username = "", password = "",
  protocol = "sftp://", port = "22")
}
\arguments{
\item{server}{This is the SFTP server URL or hostname or IP-adress.}

\item{folder}{This is the path to the folder where you want to act. The value
can be a single folder name in the SFTP root or a path with subfolders,
like 'dir1/dir2/dir3'. Note that directory names are CaSe SeNsItIvE. Default
is empty string, which takes you to the root folder.}

\item{username}{The SFTP account username. It is recommended to not store
your username in any script, but instead create a System Environment Variable
and call it using \code{Sys.getenv}}

\item{password}{The SFTP account password. It is recommended to not store
your password in any script, but instead create a System Environment Variable
and call it using \code{Sys.getenv}}

\item{protocol}{Default is 'sftp://'. You will probably never change this value.}

\item{port}{Default is '22'. You will probably never change this value. Theoretically,
an SFTP server could be set up to respond on a port other than the default port 22.}
}
\value{
the function returns a list of strings which are used by all other sftp
functions in the sftp package when connecting to an SFTP server.
}
\description{
\code{sftp_connect} outputs a list which is used as an argument value to
all other functions in the sftp package. Recommended practice is to
store the output as \code{sftp_con} since that is the object name all other
sftp functions look for by default.
}
\examples{

sftp_con <- sftp_connect(server = "sftpserver.foo",
                         folder = "myfolder",
                         username = Sys.getenv("MY_SFTP_USER"),
                         password = Sys.getenv("MY_SFTP_PASS"),
                         protocol = "sftp://",
                         port = "22")

}
\seealso{
Any other functions in the sftp package. See listing
by doing \code{?sftp}
}
