% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp_functions.R
\name{sftp_delete}
\alias{sftp_delete}
\title{Delete files in an SFTP account}
\usage{
sftp_delete(file, sftp_connection = sftp_con, verbose = TRUE,
  curlPerformVerbose = FALSE)
}
\arguments{
\item{file}{A vector of one or more file names that exist on the SFTP url.
Wildcard is not supported in order to minimize the risk of accidental deletion.
When a large portion of the files currently on the SFTP url need to be deleted, you can create
an input value for \code{file} by using \code{\link{sftp_listfiles}}.
Note! Folder (Directories) cannot be deleted. See instead \code{\link{sftp_removedir}}}

\item{sftp_connection}{A list object created by calling \code{\link{sftp_connect}}.
Default is \code{sftp_con}.}

\item{curlPerformVerbose}{Logical. Turn on messages to console form curlPerform.
Default is FALSE.}

\item{verbose{}}{Logical. Turn on messages to console. Default is TRUE}
}
\value{
The function returns the number of files deleted.
}
\description{
Specify a file name or a vector of file names to delete in an SFTP account.
}
\examples{

# minimal - delete single file and rely on defaults
sftp_delete(file = "my_bad_file.csv")

# explicit - delete several files in a vector
sftp_delete(file = files$name,
            sftp_connection = sftp_con,
            verbose = TRUE,
            curlPerformVerbose = FALSE)

}
\seealso{
\code{\link{sftp_download}}, \code{\link{sftp_upload}}, \code{\link{sftp_rename}}
}
