% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp_functions.R
\name{sftp_list}
\alias{sftp_list}
\title{List files and/or folders in an SFTP directory}
\usage{
sftp_list(sftp_connection = sftp_con, verbose = TRUE,
  curlPerformVerbose = FALSE, type)
}
\arguments{
\item{sftp_connection}{A list created by calling \code{\link{sftp_connect}}
Default value is "sftp_con".}

\item{verbose}{Logical. Turn on messages to console. Default is TRUE.}

\item{curlPerformVerbose}{Logical. Turn on messages to console from RCurl::curlPerform.
Default is FALSE.}

\item{type}{One of "f", "file" (for list files), "d", "dir", "directory" (for list directories), "all" (for both files and directories)}
}
\value{
A dataframe
}
\description{
This function lists files or folders or both in the SFTP location
specified by a list object created by calling \code{\link{sftp_connect}}.
Convenience wrapper functions are available:
\code{\link{sftp_listfiles}} and \code{\link{sftp_listdirs}}
}
\examples{

# minimal - rely on defaults
files_and_folders <- sftp_list(type = "all")

# explicit
files_and_folders <- sftp_list(sftp_connection = sftp_con,
                               verbose = TRUE,
                               curlPerformVerbose = FALSE,
                               type = "all")

}
\seealso{
\code{\link{sftp_listfiles}}, \code{\link{sftp_listdirs}}
}
