% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp_functions.R
\name{sftp_makedir}
\alias{sftp_makedir}
\title{Create a new directory in an SFTP account}
\usage{
sftp_makedir(foldername, sftp_connection = sftp_con, verbose = TRUE,
  curlPerformVerbose = FALSE)
}
\arguments{
\item{foldername}{A character vector of length 1 or more, containing the names
of folders you want to create. The folder(s) will be created below the folder
where you are currently standing according to the sftp connection list
object you are using (default 'stfp_con'). A folder with subfolder and
subsubfolder etc can be created in one single operation by supplying a path,
like "folder1/folder2/folder3".}

\item{sftp_connection}{A list object created  by calling \code{\link{sftp_connect}}.
Default is \code{sftp_con}.}

\item{verbose}{Logical. Turn on messages to console. Default is TRUE}

\item{curlPerformVerbose}{Logical. Turn on messages to console form curlPerform.
Default is FALSE.}
}
\value{
The function returns the number of new folders created.
}
\description{
Create one or several directories (subdirectories) in an SFTP account.
The function uses connection credentials from a list object
created by calling \code{\link{sftp_connect}}.
}
\examples{

# minimal - create one folder, rely on defaults
sftp_makedir("great_folder_name")

# explicit - create a multi-level path
sftp_makedir(foldername = "level1/folder2/directory3",
             sftp_connection = sftp_con,
             verbose = TRUE,
             curlPerformVerbose = FALSE)

 # create new folders as specified in a vector with names
 sftp_makedir(newfolders$names)

}
\seealso{
\code{\link{sftp_removedir}}, \code{\link{sftp_rename}}, \code{\link{sftp_changedir}}
}
