% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp_functions.R
\name{sftp_removedir}
\alias{sftp_removedir}
\title{Remove a new directory in an SFTP account}
\usage{
sftp_removedir(
  foldername,
  sftp_connection = sftp_con,
  verbose = TRUE,
  curlPerformVerbose = FALSE,
  curl_options = list()
)
}
\arguments{
\item{foldername}{A character vector of length 1 or more, containing the names
of folders you want to remove. The folder(s) will be looked for below the folder
where you are currently standing according to the sftp connection list
object you are using (default 'stfp_con'). A subfolder or a subsubfolder etc
can be removed in one single operation by supplying a path,
like "folder1/folder2/folder3". That will remove the last folder in the path, in
this case "folder3". NOTE! Non-empty directories cannot be removed.}

\item{sftp_connection}{A list object created  by calling \link{sftp_connect}.
Default is \code{sftp_con}.}

\item{verbose}{Logical. Turn on messages to console. Default is TRUE}

\item{curlPerformVerbose}{Logical. Turn on messages to console form curlPerform.
Default is FALSE.}

\item{curl_options}{A list of named values with names as listed by
\code{RCurl::listCurlOptions}. The values are handed to the .opts parameter
of the RCurl function called by \code{sftp_removedir}. Optional. Default is an empty list.}
}
\value{
The function returns the number folders deleted.
}
\description{
Remove one or several directories (subdirectories) in an SFTP account.
The function uses connection credentials from a list object
created by calling \link{sftp_connect}.
\code{sftp_removedir} internally calls \code{RCurl::curlPerform}
Questions? \href{https://github.com/stenevang/sftp}{https://github.com/stenevang/sftp}
}
\examples{

# minimal - remove one folder, rely on defaults
sftp_removedir("this_folder_is_bad")

# explicit - remove a directory at the end of a multi-level path
sftp_removedir(foldername = "level1/folder2/directory3",
             sftp_connection = sftp_con,
             verbose = TRUE,
             curlPerformVerbose = FALSE)

# remove several folders as specified in a vector of names
sftp_removedir(removefolders$names)

}
\seealso{
\link{sftp_makedir}, \link{sftp_rename}, \link{sftp_changedir}, \link{sftp}
}
