% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp_functions.R
\name{sftp_upload}
\alias{sftp_upload}
\title{Upload files to an SFTP account}
\usage{
sftp_upload(file, fromfolder = getwd(), sftp_connection = sftp_con,
  verbose = TRUE)
}
\arguments{
\item{file}{A character vector of file names. If the wildcard "*" is used, then
all files in \code{fromfolder} will be uploaded. NOTE! Names of folders (directories)
cannot be part of \code{file}. To create folders on the SFTP server, see \code{\link{sftp_makedir}}}

\item{fromfolder}{A string containing a valid path - relative or absolute - to
the folder where the files in \code{file} can be found. Default is the working directory.}

\item{sftp_connection}{A list object created with \code{link{sftp_connect}}. Default is \code{sftp_con}.}

\item{verbose}{Logical. Turn on messages to console. Default is TRUE.}
}
\value{
The function returns the number of files uploaded, following successful upload.
}
\description{
This function is used for uploading files to an SFTP account.
The function uses connection credentials from a list object
created by calling \code{\link{sftp_connect}}. The files will be uploaded to
the SFTP account folder where you are currently "standing" as specified in the
connection list object.
}
\examples{

# minimal - take all files in the current working directory
sftp_upload(file = "*")

# take a vector of file names from a dataframe
sftp_upload(file = files$name, fromfolder = "my/relative/path/")

# take one specific file name from a vector from a dataframe
sftp_upload(file = files$name[1], fromfolder = getwd())

# take all files that exist in fromfolder by using wildcard
sftp_upload(file = "*", fromfolder = getwd())


}
\seealso{
\code{\link{sftp_download}}, \code{\link{sftp_delete}}, \code{\link{sftp_rename}}
}
