% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp_functions.R
\name{trim_slashes}
\alias{trim_slashes}
\title{SFTP operations tools}
\usage{
trim_slashes(string)
}
\description{
Questions? \href{https://github.com/stenevang/sftp}{https://github.com/stenevang/sftp}
}
\details{
\link{sftp} is a function that loads the help documentation you are reading now.

_____________________________________________________________________________

\bold{CONNECT:}

\link{sftp_connect}

_____________________________________________________________________________

\bold{TRANSFER:}

\link{sftp_upload} \cr
\link{sftp_download}

_____________________________________________________________________________

\bold{NAVIGATE:}

\link{sftp_list} \cr
\link{sftp_listdirs} \cr
\link{sftp_listfiles} \cr
\link{sftp_changedir}

_____________________________________________________________________________

\bold{CREATE:}

\link{sftp_makedir} \cr
\link{sftp_removedir} \cr
\link{sftp_rename} \cr
\link{sftp_delete}
}
