% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp_functions.R
\name{sftp_connect}
\alias{sftp_connect}
\title{Create a list object with SFTP connection details}
\usage{
sftp_connect(
  server = "localhost",
  folder = NULL,
  username = NULL,
  password = NULL,
  protocol = "sftp://",
  port = 22,
  timeout = 30
)
}
\arguments{
\item{server}{This is the SFTP server URL or hostname or IP-adress. Default is localhost.}

\item{folder}{This is the path to the folder where you want to operate. The value
can be a single folder name in the SFTP root or a path with subfolders,
like 'dir1/dir2/dir3'. Note that directory names are CaSe SeNsItIvE. Default
is NULL, which takes you to the root folder.}

\item{username}{The SFTP account username. It is recommended to not store
your username in any script, but instead create a System Environment Variable
and call it using \code{Sys.getenv} Default is NULL.}

\item{password}{The SFTP account password. It is recommended to not store
your password in any script, but instead create a System Environment Variable
and call it using \code{Sys.getenv} IMPORTANT! Your password MUST NOT contain
the colon character : or the at-sign @ since these characters have special use
as delimiters in the connection string! Default is NULL.}

\item{protocol}{Default is 'sftp://'. You will probably never change this value.}

\item{port}{Default is '22'. You will probably never change this value. Theoretically,
an SFTP server could be set up to respond on a port other than the default port 22.}

\item{timeout}{RCurl connection timeout in seconds. Default is 30.}
}
\value{
The function returns a list of strings which are used by all other
functions in the sftp package when connecting to an SFTP server.
}
\description{
\code{sftp_connect} outputs a list which is used as an argument to
all other functions of the sftp package. Recommended practice is to
store the output as \code{sftp_con} since that is the object name all other
sftp functions look for by default. IMPORTANT! Your password MUST NOT contain
the colon character : or the at-sign @ since these characters have special use
as delimiters in the connection string!
Questions? \href{https://github.com/stenevang/sftp}{https://github.com/stenevang/sftp}
}
\examples{

# minimal - rely on defaults
sftp_con <- sftp_connect(username = Sys.getenv("MY_SFTP_USER"),
                         password = Sys.getenv("MY_SFTP_PASS"))

# explicit
sftp_con <- sftp_connect(server = "mysftp.com",
                         folder = "myfolder",
                         username = Sys.getenv("MY_SFTP_USER"),
                         password = Sys.getenv("MY_SFTP_PASS"),
                         protocol = "sftp://",
                         port = 22,
                         timeout = 60)

}
\seealso{
See listing here: \link{sftp}
}
