% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp_functions.R
\name{sftp_delete}
\alias{sftp_delete}
\title{Delete files in an SFTP account}
\usage{
sftp_delete(
  file,
  sftp_connection = sftp_con,
  verbose = TRUE,
  curlPerformVerbose = FALSE,
  curl_options = list()
)
}
\arguments{
\item{file}{A vector of one or more file names that exist on the SFTP url.
Wildcard is not supported in order to minimize the risk of accidental deletion.
When a large portion of the files currently on the SFTP url needs to be deleted, you can create
an input value for \code{file} by using \link{sftp_listfiles}.
Note! Folders (directories) cannot be deleted. See instead \link{sftp_removedir}}

\item{sftp_connection}{A list object created by calling \link{sftp_connect}.
Default is 'sftp_con'.}

\item{verbose}{Logical. Turn on messages to console. Default is TRUE}

\item{curlPerformVerbose}{Logical. Turn on messages to console form curlPerform.
Default is FALSE.}

\item{curl_options}{A list of named values with names as listed by
\code{RCurl::listCurlOptions}. The values are handed to the .opts parameter
of the RCurl function called by \code{sftp_delete}. Optional. Default is an empty list.}
}
\value{
The function returns the number of files deleted.
}
\description{
Specify a file name or a vector of file names to delete in an SFTP account.
\code{sftp_delete} internally calls \code{RCurl::curlPerform}
Questions? \href{https://github.com/stenevang/sftp}{https://github.com/stenevang/sftp}
}
\examples{

# minimal - delete single file and rely on defaults
sftp_delete("my_bad_file.csv")

# explicit - delete several files in a vector
sftp_delete(file = files$name,
            sftp_connection = sftp_con,
            verbose = TRUE,
            curlPerformVerbose = FALSE)

}
\seealso{
\link{sftp_download}, \link{sftp_upload}, \link{sftp_rename}, \link{sftp}
}
