% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp_functions.R
\name{sftp_list}
\alias{sftp_list}
\title{List files and/or folders in an SFTP directory}
\usage{
sftp_list(
  sftp_connection = sftp_con,
  verbose = TRUE,
  curlPerformVerbose = FALSE,
  encoding = "UTF-8",
  type = "all",
  recurse = F,
  curl_options = list()
)
}
\arguments{
\item{sftp_connection}{A list created by calling \link{sftp_connect}
Default value is \code{sftp_con}.}

\item{verbose}{Logical. Turn on messages to console. Default is TRUE.}

\item{curlPerformVerbose}{Logical. Turn on messages to console from RCurl::curlPerform.
Default is FALSE.}

\item{encoding}{Explicitly identifies the encoding of the content that is returned by the server in its response. Possible values are ‘UTF-8’ or ‘ISO-8859-1’. Default is 'UTF-8'.}

\item{type}{One of "f", "file" (for list files), "d", "dir", "directory" (for list directories), "all" (for both files and directories, default).}

\item{recurse}{Logical. Recurse directories? Default is FALSE.}

\item{curl_options}{A list of named values with names as listed by \code{RCurl::listCurlOptions}. The values are handed to the .opts parameter of the RCurl function called by \code{sftp_list}. Optional. Default is an empty list.}
}
\value{
A dataframe
}
\description{
This function lists files, or folders, or both in the SFTP location
specified by a list object created by calling \link{sftp_connect}.
\code{sftp_list} internally calls \code{RCurl::getURL}
Convenience wrapper functions are available:
\link{sftp_listfiles} and \link{sftp_listdirs}
Questions? \href{https://github.com/stenevang/sftp}{https://github.com/stenevang/sftp}
}
\examples{

# minimal - rely on defaults
files_and_folders <- sftp_list(type = "all")

# explicit
files_and_folders <- sftp_list(sftp_connection = sftp_con,
                               verbose = TRUE,
                               curlPerformVerbose = FALSE,
                               type = "all")

}
\seealso{
\link{sftp_listfiles}, \link{sftp_listdirs}, \link{sftp}
}
