% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp_functions.R
\name{sftp_listdirs}
\alias{sftp_listdirs}
\title{List folders in an SFTP directory}
\usage{
sftp_listdirs(sftp_connection = sftp_con, verbose = TRUE,
  curlPerformVerbose = FALSE, recurse = FALSE)
}
\arguments{
\item{sftp_connection}{A list created by calling \link{sftp_connect}
Default value is \code{sftp_con}.}

\item{verbose}{Logical. Turn on messages to console. Default is TRUE.}

\item{curlPerformVerbose}{Logical. Turn on messages to console from RCurl::curlPerform.
Default is FALSE.}

\item{recurse}{Logical. Recurse directories? Default is FALSE.}
}
\value{
A dataframe
}
\description{
Convenience function wrapping \link{sftp_list} with type hard set to "dir".
Questions? \href{https://github.com/stenevang/sftp}{https://github.com/stenevang/sftp}
}
\examples{
# minimal - rely on defaults
files <- sftp_listdirs()

# explicit
files <- sftp_listdirs(sftp_connection = sftp_con,
                       verbose = TRUE,
                       curlPerformVerbose = FALSE)

}
\seealso{
\link{sftp_listfiles}, \link{sftp_list}, \link{sftp}
}
