% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp_functions.R
\name{sftp_rename}
\alias{sftp_rename}
\title{Rename a single file or single directory in an SFTP account}
\usage{
sftp_rename(from, to, sftp_connection = sftp_con, verbose = TRUE,
  curlPerformVerbose = FALSE)
}
\arguments{
\item{from}{A single name of an existing file or folder. If a vector of names is supplied, only
the first value will be used, and there will be a warning.}

\item{to}{A single name of an existing file or folder. If a vector of names is supplied, only
the first value will be used, and there will be a warning.  #'}

\item{sftp_connection}{A list object created by calling \code{\link{sftp_connect}}.
Default is \code{sftp_con}.}

\item{verbose}{Logical. Turn on messages to console. Default is TRUE}

\item{curlPerformVerbose}{Logical. Turn on messages to console form curlPerform.
Default is FALSE.}
}
\value{
The function returns TRUE if successful.
}
\description{
Specify a current name and a new name, for a file or folder existing in an SFTP account.
While folders must be empty when using \code{\link{sftp_removedir}}, that is not the case with
\code{sftp_rename} - non-empty folders can be renamed.
}
\examples{

# minimal - rename single file, rely on defaults
sftp_rename(from = "oldname.csv",
            to = "newname.csv")

# explicit - rename a folder
sftp_rename(from = "old_folder_name",
            to = "new_folder_name",
            sftp_connection = sftp_con,
            verbose = TRUE,
            curlPerformVerbose = FALSE)

}
\seealso{
\code{\link{sftp_download}}, \code{\link{sftp_upload}}, \code{\link{sftp_delete}}
}
